requirements.txt
anyio==4.0.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pip==23.2.1
python-telegram-bot==20.5
requests==2.31.0
setuptools==68.0.0
sniffio==1.3.0
soupsieve==2.5
urllib3==2.0.4
wheel==0.41.2
